Inspector/Analyst refactoring
