+++ java-tools/OsmAndServer/src/main/java/net/osmand/server/api/searchtest/OBFService.java
@@ -31,6 +31,55 @@ public interface OBFService extends BaseService {
 				}
 			});
 
+	Map<String, Map<Integer, InspectorService.ObfFieldSpec>> OBF_MESSAGE_SCHEMA = buildObfMessageSchema();
+
+	Map<Integer, String> OBF_STRUCTURE_FIELD_NAMES = buildObfStructureFieldNames();
+
+	record ObfFieldSpec(String fieldName, String childMessageType, InspectorService.ObfLengthType lengthType, boolean packedVarInt, boolean repeated) {}
+
+	static String toLowerCamelFromUpperUnderscore(String upperUnderscore) {
+		String[] parts = upperUnderscore.toLowerCase(Locale.ROOT).split("_");
+		if (parts.length == 0) {
+			return upperUnderscore;
+		}
+		StringBuilder sb = new StringBuilder(parts[0]);
+		for (int i = 1; i < parts.length; i++) {
+			String p = parts[i];
+			if (p.isEmpty()) {
+				continue;
+			}
+			sb.append(Character.toUpperCase(p.charAt(0)));
+			if (p.length() > 1) {
+				sb.append(p.substring(1));
+			}
+		}
+		return sb.toString();
+	}
+	
+	static Map<Integer, String> buildObfStructureFieldNames() {
+		Map<Integer, String> out = new HashMap<>();
+		try {
+			for (java.lang.reflect.Field f : OsmandOdb.OsmAndStructure.class.getDeclaredFields()) {
+				if (!java.lang.reflect.Modifier.isStatic(f.getModifiers())) {
+					continue;
+				}
+				if (f.getType() != int.class) {
+					continue;
+				}
+				String name = f.getName();
+				if (!name.endsWith("_FIELD_NUMBER")) {
+					continue;
+				}
+				int fieldNumber = f.getInt(null);
+				String base = name.substring(0, name.length() - "_FIELD_NUMBER".length());
+				String fieldName = toLowerCamelFromUpperUnderscore(base);
+				out.put(fieldNumber, fieldName);
+			}
+		} catch (IllegalAccessException e) {
+			throw new RuntimeException(e);
+		}
+		return out;
+	}
 	OsmAndMapsService getMapsService();
 	String getSearchTestDatasourceUrl();
 
@@ -59,6 +108,16 @@ public interface OBFService extends BaseService {
 	}
 
 	record CachedIndexTokens(String cacheKey, long fileLength, long lastModified, List<IndexToken> tokens) {}
+
+    class BinaryMapIndexReaderExt extends BinaryMapIndexReader {
+        BinaryMapIndexReaderExt(final RandomAccessFile raf, File file) throws IOException {
+            super(raf, file);
+        }
+
+        CodedInputStream getInputStream() {
+            return codedIS;
+        }
+    }
 
 	List<String> OBF_CONTINENTS = List.of(
 			"asia",
@@ -684,5 +743,127 @@ public interface OBFService extends BaseService {
 		}
 		String normalizedValue = Normalizer.normalize(value, Normalizer.Form.NFKC).toLowerCase(Locale.ROOT);
 		return normalizedValue.replace("\u0307", "");
+	}
+
+	static Map<String, Map<Integer, InspectorService.ObfFieldSpec>> buildObfMessageSchema() {
+		Map<String, Map<Integer, InspectorService.ObfFieldSpec>> schema = new HashMap<>();
+		addObfSpec(schema, "OsmAndStructure", 7, "addressIndex", "OsmAndAddressIndex", InspectorService.ObfLengthType.FIXED32, false, true);
+		addObfSpec(schema, "OsmAndStructure", 4, "transportIndex", "OsmAndTransportIndex", InspectorService.ObfLengthType.FIXED32, false, true);
+		addObfSpec(schema, "OsmAndStructure", 8, "poiIndex", "OsmAndPoiIndex", InspectorService.ObfLengthType.FIXED32, false, true);
+		addObfSpec(schema, "OsmAndStructure", 6, "mapIndex", "OsmAndMapIndex", InspectorService.ObfLengthType.FIXED32, false, true);
+		addObfSpec(schema, "OsmAndStructure", 9, "routingIndex", "OsmAndRoutingIndex", InspectorService.ObfLengthType.FIXED32, false, true);
+		addObfSpec(schema, "OsmAndStructure", 10, "hhRoutingIndex", "OsmAndHHRoutingIndex", InspectorService.ObfLengthType.FIXED32, false, true);
+
+		addObfSpec(schema, "OsmAndTileBox", 1, "left", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndTileBox", 2, "right", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndTileBox", 3, "top", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndTileBox", 4, "bottom", null, InspectorService.ObfLengthType.VAR_INT);
+
+		addObfSpec(schema, "OsmAndPoiIndex", 1, "name", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiIndex", 2, "boundaries", "OsmAndTileBox", InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiIndex", 3, "categoriesTable", "OsmAndCategoryTable", InspectorService.ObfLengthType.VAR_INT, false, true);
+		addObfSpec(schema, "OsmAndPoiIndex", 4, "nameIndex", "OsmAndPoiNameIndex", InspectorService.ObfLengthType.FIXED32);
+		addObfSpec(schema, "OsmAndPoiIndex", 5, "subtypesTable", "OsmAndSubtypesTable", InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiIndex", 6, "boxes", "OsmAndPoiBox", InspectorService.ObfLengthType.FIXED32, false, true);
+		addObfSpec(schema, "OsmAndPoiIndex", 9, "poiData", "OsmAndPoiBoxData", InspectorService.ObfLengthType.FIXED32, false, true);
+		addObfSpec(schema, "OsmAndSubtypesTable", 4, "subtypes", "OsmAndPoiSubtype", InspectorService.ObfLengthType.VAR_INT, false, true);
+		addObfSpec(schema, "OsmAndCategoryTable", 1, "category", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndCategoryTable", 3, "subcategories", null, InspectorService.ObfLengthType.VAR_INT);
+
+		addObfSpec(schema, "OsmAndPoiNameIndex", 3, "table", "IndexedStringTable", InspectorService.ObfLengthType.FIXED32);
+		addObfSpec(schema, "OsmAndPoiNameIndex", 5, "data", "OsmAndPoiNameIndexData", InspectorService.ObfLengthType.VAR_INT, false, true);
+		addObfSpec(schema, "OsmAndPoiNameIndexData", 3, "atoms", "OsmAndPoiNameIndexDataAtom", InspectorService.ObfLengthType.VAR_INT, false, true);
+		addObfSpec(schema, "OsmAndPoiNameIndexDataAtom", 2, "zoom", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiNameIndexDataAtom", 3, "x", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiNameIndexDataAtom", 4, "y", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiNameIndexDataAtom", 5, "bloomIndex", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiNameIndexDataAtom", 14, "shiftTo", null, InspectorService.ObfLengthType.FIXED32);
+
+		addObfSpec(schema, "IndexedStringTable", 1, "prefix", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "IndexedStringTable", 3, "key", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "IndexedStringTable", 4, "val", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "IndexedStringTable", 5, "subtables", "IndexedStringTable", InspectorService.ObfLengthType.VAR_INT, false, true);
+
+		addObfSpec(schema, "OsmAndPoiBox", 1, "zoom", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiBox", 2, "left", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiBox", 3, "top", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiBox", 4, "categories", "OsmAndPoiCategories", InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiBox", 8, "tagGroups", "OsmAndPoiTagGroups", InspectorService.ObfLengthType.FIXED32);
+		addObfSpec(schema, "OsmAndPoiBox", 10, "subBoxes", "OsmAndPoiBox", InspectorService.ObfLengthType.FIXED32, false, true);
+		addObfSpec(schema, "OsmAndPoiBox", 14, "shiftToData", null, InspectorService.ObfLengthType.FIXED32);
+		addObfPackedVarIntSpec(schema, "OsmAndPoiCategories", 3, "categories");
+		addObfPackedVarIntSpec(schema, "OsmAndPoiCategories", 5, "subcategories");
+		addObfPackedVarIntSpec(schema, "OsmAndPoiTagGroups", 2, "ids");
+		addObfSpec(schema, "OsmAndPoiTagGroups", 5, "groups", "OsmAndPoiTagGroup", InspectorService.ObfLengthType.VAR_INT, false, true);
+		addObfSpec(schema, "OsmAndPoiBoxData", 5, "poiData", "OsmAndPoiBoxDataAtom", InspectorService.ObfLengthType.VAR_INT, false, true);
+		addObfSpec(schema, "OsmAndPoiTagGroup", 1, "id", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiTagGroup", 5, "tagValues", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiSubtype", 1, "name", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiSubtype", 2, "tagname", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiSubtype", 3, "isText", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiSubtype", 5, "frequency", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiSubtype", 6, "subtypeValuesSize", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiSubtype", 8, "subtypeValue", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiBoxData", 1, "zoom", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiBoxData", 2, "x", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiBoxData", 3, "y", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiBoxDataAtom", 2, "dx", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiBoxDataAtom", 3, "dy", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfPackedVarIntSpec(schema, "OsmAndPoiBoxDataAtom", 4, "categories");
+		addObfPackedVarIntSpec(schema, "OsmAndPoiBoxDataAtom", 5, "subcategories");
+		addObfSpec(schema, "OsmAndPoiBoxDataAtom", 6, "name", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiBoxDataAtom", 7, "nameEn", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiBoxDataAtom", 8, "id", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiBoxDataAtom", 10, "openingHours", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiBoxDataAtom", 11, "site", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiBoxDataAtom", 12, "phone", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiBoxDataAtom", 13, "note", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfPackedVarIntSpec(schema, "OsmAndPoiBoxDataAtom", 14, "textCategories");
+		addObfSpec(schema, "OsmAndPoiBoxDataAtom", 15, "textValues", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndPoiBoxDataAtom", 16, "precisionXY", null, InspectorService.ObfLengthType.VAR_INT);
+		addObfPackedVarIntSpec(schema, "OsmAndPoiBoxDataAtom", 17, "tagGroups");
+
+		addObfSpec(schema, "OsmAndAddressIndex", 3, "boundaries", "OsmAndTileBox", InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndAddressIndex", 4, "attributeTagsTable", "StringTable", InspectorService.ObfLengthType.VAR_INT);
+		addObfSpec(schema, "OsmAndAddressIndex", 6, "cities", "CitiesIndex", InspectorService.ObfLengthType.FIXED32, false, true);
+		addObfSpec(schema, "OsmAndAddressIndex", 7, "nameIndex", "OsmAndAddressNameIndexData", InspectorService.ObfLengthType.FIXED32);
+
+		addObfSpec(schema, "CitiesIndex", 5, "cities", "CityIndex", InspectorService.ObfLengthType.VAR_INT, false, true);
+		addObfSpec(schema, "CitiesIndex", 7, "blocks", "CityBlockIndex", InspectorService.ObfLengthType.VAR_INT, false, true);
+		addObfSpec(schema, "CityBlockIndex", 10, "buildings", "BuildingIndex", InspectorService.ObfLengthType.VAR_INT, false, true);
+		addObfSpec(schema, "CityBlockIndex", 12, "streets", "StreetIndex", InspectorService.ObfLengthType.VAR_INT, false, true);
+		addObfSpec(schema, "StreetIndex", 5, "intersections", "StreetIntersection", InspectorService.ObfLengthType.VAR_INT, false, true);
+		addObfSpec(schema, "StreetIndex", 12, "buildings", "BuildingIndex", InspectorService.ObfLengthType.VAR_INT, false, true);
+
+		addObfSpec(schema, "OsmAndAddressNameIndexData", 4, "table", "IndexedStringTable", InspectorService.ObfLengthType.FIXED32);
+		addObfSpec(schema, "OsmAndAddressNameIndexData", 7, "atom", "AddressNameIndexData", InspectorService.ObfLengthType.VAR_INT, false, true);
+		addObfSpec(schema, "AddressNameIndexData", 4, "atom", "AddressNameIndexDataAtom", InspectorService.ObfLengthType.VAR_INT, false, true);
+
+		addObfSpec(schema, "OsmAndMapIndex", 4, "rules", "OsmAndMapIndex.MapEncodingRule", InspectorService.ObfLengthType.VAR_INT, false, true);
+		addObfSpec(schema, "OsmAndMapIndex", 5, "levels", "OsmAndMapIndex.MapRootLevel", InspectorService.ObfLengthType.FIXED32, false, true);
+		addObfSpec(schema, "OsmAndMapIndex.MapRootLevel", 7, "boxes", "OsmAndMapIndex.MapDataBox", InspectorService.ObfLengthType.VAR_INT, false, true);
+		addObfSpec(schema, "OsmAndMapIndex.MapRootLevel", 15, "blocks", "MapDataBlock", InspectorService.ObfLengthType.VAR_INT, false, true);
+		addObfSpec(schema, "OsmAndMapIndex.MapDataBox", 7, "boxes", "OsmAndMapIndex.MapDataBox", InspectorService.ObfLengthType.VAR_INT, false, true);
+		addObfSpec(schema, "MapDataBlock", 12, "dataObjects", "MapData", InspectorService.ObfLengthType.VAR_INT, false, true);
+		addObfSpec(schema, "MapDataBlock", 15, "stringTable", "StringTable", InspectorService.ObfLengthType.VAR_INT);
+
+		return schema;
+	}
+
+	static void addObfPackedVarIntSpec(Map<String, Map<Integer, InspectorService.ObfFieldSpec>> schema, String messageType,
+	                                   int fieldNumber, String fieldName) {
+		addObfSpec(schema, messageType, fieldNumber, fieldName, null, InspectorService.ObfLengthType.VAR_INT, true, false);
+	}
+
+	static void addObfSpec(Map<String, Map<Integer, InspectorService.ObfFieldSpec>> schema, String messageType,
+	                       int fieldNumber, String fieldName, String childMessageType, InspectorService.ObfLengthType lengthType) {
+		addObfSpec(schema, messageType, fieldNumber, fieldName, childMessageType, lengthType, false, false);
+	}
+
+	static void addObfSpec(Map<String, Map<Integer, InspectorService.ObfFieldSpec>> schema, String messageType,
+	                       int fieldNumber, String fieldName, String childMessageType, InspectorService.ObfLengthType lengthType,
+	                       boolean packedVarInt, boolean repeated) {
+		Map<Integer, InspectorService.ObfFieldSpec> byField = schema.computeIfAbsent(messageType, k -> new HashMap<>());
+		byField.put(fieldNumber, new InspectorService.ObfFieldSpec(fieldName, childMessageType, lengthType, packedVarInt, repeated));
 	}
 }
