+++ java-tools/OsmAndServer/src/main/java/net/osmand/server/api/services/SearchTestService.java
@@ -41,7 +41,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 
 @Service
-public class SearchTestService implements ReportService, DataService, DetectorService, InspectorService, AnalystService {
+public class SearchTestService implements ReportService, DataService, DetectorService, InspectorService, AnalystService, TokenAnalystService, AddressPOIAnalystService {
     /**
      * Lightweight DTO for listing test-cases with parent dataset name.
      */
